Refactor useTextField
